refactor: Refactor the folder structure
