--- app/global/constants.ts
@@ -1,6 +1,0 @@
-const constants = {
-  // Timing
-  carouselTiming: 3900,
-};
-
-export default constants;
