+++ src/components/index.ts
@@ -1,0 +1,5 @@
+import LoginWithGoogleButton from './global/LoginWithGoogleButton';
+import Carousel from './home/carousel';
+import Section from './home/section';
+
+export { Carousel, LoginWithGoogleButton, Section };
