+++ src/global/constants.ts
@@ -1,0 +1,6 @@
+const CONSTANTS = {
+  // Timing
+  CAROUSEL_SLIDER_TRANSITION_TIME: 3900,
+};
+
+export default CONSTANTS;
