+++ src/global/index.ts
@@ -1,0 +1,4 @@
+import colors from './colors';
+import CONSTANTS from './constants';
+
+export { CONSTANTS, colors };
